requirements.txt
certifi==2018.1.18
chardet==3.0.4
feedparser==5.2.1
idna==2.6
LEPL==5.1.3
lxml==4.1.1
PyYAML==3.12
requests==2.18.4
rfc6266==0.0.4
urllib3==1.22
